Addition methods, GUI actualized, classes actualized.
We implemented the methods to add an object from "RestaurantManagerGUI" to the data of "RestaurantManager" with dependency only between "RestaurantManagerGUI" and "RestaurantManager". Also actualized the constructor of some classes to take the id as a long , and not as a int. Finally, the window to create an user has been actualized with two more text area to put the name and last name of the employee who owns the new user.