Login, GUI improvement.
We implemented the method "login" to give check if an user it's registered or not. Also implement the "show" methods to switch between the "create" windows.  Also improve the GUI of the main pane to always show the active user in the lower left corner. Finally, we implemented the method to initialize the table views of Ingredient window.